GetFood script with Feature for clicking and dragging Items with the tag of food or HpUp, issues in description
In the test project I am using for the code, there is an offset between where the mouse is and where the object is. Unsure how to fix this. There is also a jump to a certain point when clicking the object, likely due to the conversion from pixels to coordinates